fix/#48: monWord 페이지 로직 변경
