+++ src/app/(main)/MonWord/components/WordList.tsx
@@ -37,7 +37,7 @@ export default function WordList() {
       const data = await monWordApi.postMonWordDone();
       // console.log(data);
       alert("오늘의 MonWord를 완료했어요!");
-      router.push("/");
+      router.push("/student");
     } catch (error: any) {
       if (error.response) {
         alert(error.response.data.message);
@@ -50,26 +50,6 @@ export default function WordList() {
     }
   };
 
-  // mon단어 item 이해했어요 post
-  const postTodayMonWordItemUnderstand = async (wordId: number) => {
-    try {
-      setIsLoading(true);
-      const data = await monWordApi.postWordItemUnderstand(wordId);
-      console.log(data);
-    } catch (error: any) {
-      if (error.response) {
-        alert(error.response.data.message);
-      } else {
-        alert("알 수 없는 오류가 발생했습니다.");
-      }
-      console.error("오늘의 monWord 학습 완료 post 실패", error);
-    }
-  };
-
-  const handleFinishClick = () => {
-    postTodayMonWordDone();
-  };
-
   return (
     <>
       <div className="px-5 py-7 flex flex-col justify-center items-center gap-5">
@@ -80,14 +60,13 @@ export default function WordList() {
             word={item.word}
             explain={item.explain}
             use={item.use}
-            understandFunc={() => postTodayMonWordItemUnderstand(item.id)}
           />
         ))}
         <div className="">
           <CommonBtn
             type="finish"
             subText="오늘의 MON단어를 다 이해했어요!"
-            onClick={handleFinishClick}
+            onClick={postTodayMonWordDone}
           />
         </div>
       </div>
